renaming cln file, class and get_version
